Conflicts:
	src/crypto/ms_srtp.c

+++ src/crypto/ms_srtp.c
@@ -324,12 +324,12 @@ static int ms_add_srtp_stream(srtp_t srtp, MSCryptoSuite suite, uint32_t ssrc, c
 	/* check if key length match given policy */
 	if (stream_type==MSSRTP_ALL_STREAMS || stream_type==MSSRTP_RTP_STREAM) {
 		if (key_length != policy.rtp.cipher_key_len) {
-			ms_error("Key size (%ld) doesn't match the selected srtp profile (required %d)", key_length, policy.rtp.cipher_key_len);	
+			ms_error("Key size (%i) doesn't match the selected srtp profile (required %d)", (int)key_length, policy.rtp.cipher_key_len);	
 			return -1;
 		}
 	} else {
 		if (key_length != policy.rtcp.cipher_key_len) {
-			ms_error("Key size (%ld) doesn't match the selected srtp profile (required %d)", key_length, policy.rtcp.cipher_key_len);	
+			ms_error("Key size (%i) doesn't match the selected srtp profile (required %d)", (int)key_length, policy.rtcp.cipher_key_len);	
 			return -1;
 		}
 	}
